requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
selenium>=4.10.0
beautifulsoup4>=4.12.2
openai>=1.3.0
numpy>=1.26.0
pandas>=2.1.4
msal>=1.24.1
oura>=1.3.0
pytest>=7.4.0
black>=23.3.0
python-dateutil>=2.8.2
markdown>=3.5.2
weasyprint>=60.1
setuptools>=69.0.2
